requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.1
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
webrtcvad==2.0.10
xxhash==3.5.0
yarl==1.20.0
